// src/interfaces/ElementProps.tsx
export interface ElementProps {
  elementName: string;
  attributes: string;
  dialogTitle: string;
  aNumber: string;
  zNumber: string;
  nNumber: string;
  eNumber: string;
  uNumber: string;
  mainFeatures: string;
  eletronicConfig: string;
  discoveredBy: string;
  discoveredIn: string;
  howIsItFound: string;
  howIsItExtracted: string;
  isThereAbundance: string;
  biggestWorldReserves: string;
  howIsItUse: string;
  howShouldItBeManaged: string;
  offersRisk: string;
}